# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Translate Arxiv Paper", "version": "0.1.0"}, "paths": {"/api/v1/translate/arxiv/{arxiv_paper_id}": {"post": {"summary": "Translate Sync", "operationId": "translate_sync_api_v1_translate_arxiv__arxiv_paper_id__post", "parameters": [{"name": "arxiv_paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The ID of the paper", "title": "Arxiv Paper Id"}, "description": "The ID of the paper"}, {"name": "target_language", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/TargetLanguage", "description": "The target language"}, "description": "The target language"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TranslateResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/translate/arxiv/{arxiv_paper_id}/stream": {"get": {"summary": "Translate Stream", "operationId": "translate_stream_api_v1_translate_arxiv__arxiv_paper_id__stream_get", "parameters": [{"name": "arxiv_paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The ID of the paper", "title": "Arxiv Paper Id"}, "description": "The ID of the paper"}, {"name": "target_language", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/TargetLanguage", "description": "The target language"}, "description": "The target language"}], "responses": {"200": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/": {"get": {"summary": "List Blogs", "operationId": "list_blogs_api_v1_blog__get", "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "description": "Page number", "default": 1, "title": "Page"}, "description": "Page number"}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Items per page", "default": 10, "title": "Page Size"}, "description": "Items per page"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedBlogPostResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/my": {"get": {"summary": "List My Blogs", "operationId": "list_my_blogs_api_v1_blog_my_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "description": "Page number", "default": 1, "title": "Page"}, "description": "Page number"}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Items per page", "default": 10, "title": "Page Size"}, "description": "Items per page"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedBlogPostResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/my/generation-count": {"get": {"summary": "Get My Generation Count", "operationId": "get_my_generation_count_api_v1_blog_my_generation_count_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerationCountResponseSchema"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/blog/arxiv/{arxiv_paper_id}": {"post": {"summary": "Generate Blog", "operationId": "generate_blog_api_v1_blog_arxiv__arxiv_paper_id__post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "arxiv_paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The arXiv paper ID", "title": "Arxiv Paper Id"}, "description": "The arXiv paper ID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlogPostResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/arxiv/{arxiv_paper_id}/stream": {"get": {"summary": "Generate Blog Stream", "operationId": "generate_blog_stream_api_v1_blog_arxiv__arxiv_paper_id__stream_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "arxiv_paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The arXiv paper ID", "title": "Arxiv Paper Id"}, "description": "The arXiv paper ID"}], "responses": {"200": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/{paper_id}/rag/text": {"post": {"summary": "Rag Search Text", "operationId": "rag_search_text_api_v1_blog__paper_id__rag_text_post", "parameters": [{"name": "paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The paper ID", "title": "Paper Id"}, "description": "The paper ID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RagSearchRequestSchema"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RagSearchTextResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/{paper_id}/rag/image": {"post": {"summary": "Rag Search Image", "operationId": "rag_search_image_api_v1_blog__paper_id__rag_image_post", "parameters": [{"name": "paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The paper ID", "title": "Paper Id"}, "description": "The paper ID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RagSearchRequestSchema"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RagSearchImageResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/{paper_id}": {"get": {"summary": "Get Blog", "operationId": "get_blog_api_v1_blog__paper_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The paper ID", "title": "Paper Id"}, "description": "The paper ID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlogPostResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/blog/pdf": {"post": {"summary": "Generate Blog From Pdf", "operationId": "generate_blog_from_pdf_api_v1_blog_pdf_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_blog_from_pdf_api_v1_blog_pdf_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlogPostResponseSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/blog/pdf/stream": {"post": {"summary": "Generate Blog From Pdf Stream", "operationId": "generate_blog_from_pdf_stream_api_v1_blog_pdf_stream_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_blog_from_pdf_stream_api_v1_blog_pdf_stream_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/chat/paper/{paper_id}": {"post": {"summary": "Chat with a paper using Gemini (SSE stream)", "operationId": "chat_with_paper_api_v1_chat_paper__paper_id__post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "paper_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Paper Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}}, "responses": {"200": {"description": "SSE stream of ChatStreamEvent objects", "content": {"text/event-stream": {}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/chat/paper/{paper_id}/threads": {"get": {"summary": "List chat threads for the current user on a paper", "operationId": "list_chat_threads_api_v1_chat_paper__paper_id__threads_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "paper_id", "in": "path", "required": true, "schema": {"type": "string", "description": "The paper ID", "title": "Paper Id"}, "description": "The paper ID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatThreadListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/chat/threads/{thread_id}": {"get": {"summary": "Get full message history for a chat thread", "operationId": "get_chat_thread_api_v1_chat_threads__thread_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "thread_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "description": "The thread ID", "title": "Thread Id"}, "description": "The thread ID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatThreadResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete a chat thread owned by the current user", "operationId": "delete_chat_thread_api_v1_chat_threads__thread_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "thread_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "description": "The thread ID", "title": "Thread Id"}, "description": "The thread ID"}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/chat/my/daily-count": {"get": {"summary": "Get the current user's chat usage for today", "operationId": "get_my_chat_daily_count_api_v1_chat_my_daily_count_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatDailyCountResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/billing/me": {"get": {"summary": "Get My Subscription", "operationId": "get_my_subscription_api_v1_billing_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MySubscriptionResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/billing/checkout-session": {"post": {"summary": "Create Checkout Session", "operationId": "create_checkout_session_api_v1_billing_checkout_session_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckoutSessionResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/billing/portal-session": {"post": {"summary": "Create Portal Session", "operationId": "create_portal_session_api_v1_billing_portal_session_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckoutSessionResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/billing/webhook": {"post": {"summary": "Stripe Webhook", "operationId": "stripe_webhook_api_v1_billing_webhook_post", "parameters": [{"name": "stripe-signature", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Stripe-Signature"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/sentry-debug": {"get": {"summary": "Trigger Error", "operationId": "trigger_error_sentry_debug_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"BlogPostResponseSchema": {"properties": {"paper_id": {"type": "string", "title": "Paper Id"}, "title": {"type": "string", "title": "Title"}, "summary": {"type": "string", "title": "Summary"}, "authors": {"items": {"type": "string"}, "type": "array", "title": "Authors"}, "source_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Source Url"}, "content": {"type": "string", "title": "Content"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["paper_id", "title", "summary", "authors", "source_url", "content", "created_at", "updated_at"], "title": "BlogPostResponseSchema"}, "Body_generate_blog_from_pdf_api_v1_blog_pdf_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File", "description": "PDF file of the paper"}}, "type": "object", "required": ["file"], "title": "Body_generate_blog_from_pdf_api_v1_blog_pdf_post"}, "Body_generate_blog_from_pdf_stream_api_v1_blog_pdf_stream_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File", "description": "PDF file of the paper"}}, "type": "object", "required": ["file"], "title": "Body_generate_blog_from_pdf_stream_api_v1_blog_pdf_stream_post"}, "ChatDailyCountResponse": {"properties": {"daily": {"type": "integer", "title": "Daily"}, "limit": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Limit"}}, "type": "object", "required": ["daily", "limit"], "title": "ChatDailyCountResponse"}, "ChatMessageResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "role": {"type": "string", "enum": ["user", "assistant"], "title": "Role"}, "content": {"items": {"oneOf": [{"$ref": "#/components/schemas/TextBlock"}, {"$ref": "#/components/schemas/ToolUseBlock"}, {"$ref": "#/components/schemas/ToolResultBlock"}], "discriminator": {"propertyName": "type", "mapping": {"text": "#/components/schemas/TextBlock", "tool_result": "#/components/schemas/ToolResultBlock", "tool_use": "#/components/schemas/ToolUseBlock"}}}, "type": "array", "title": "Content"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}}, "type": "object", "required": ["id", "role", "content", "timestamp"], "title": "ChatMessageResponse"}, "ChatRequest": {"properties": {"message": {"type": "string", "title": "Message", "description": "User message text"}, "thread_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Thread Id", "description": "Existing thread ID to continue"}}, "type": "object", "required": ["message"], "title": "ChatRequest"}, "ChatThreadListResponse": {"properties": {"threads": {"items": {"$ref": "#/components/schemas/ChatThreadSummaryResponse"}, "type": "array", "title": "Threads"}}, "type": "object", "required": ["threads"], "title": "ChatThreadListResponse"}, "ChatThreadResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "paper_id": {"type": "string", "title": "Paper Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "messages": {"items": {"$ref": "#/components/schemas/ChatMessageResponse"}, "type": "array", "title": "Messages"}}, "type": "object", "required": ["id", "paper_id", "created_at", "updated_at"], "title": "ChatThreadResponse"}, "ChatThreadSummaryResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "paper_id": {"type": "string", "title": "Paper Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "message_count": {"type": "integer", "title": "Message Count"}, "title": {"type": "string", "title": "Title"}}, "type": "object", "required": ["id", "paper_id", "created_at", "updated_at", "message_count", "title"], "title": "ChatThreadSummaryResponse"}, "CheckoutSessionResponse": {"properties": {"url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["url"], "title": "CheckoutSessionResponse"}, "GenerationCountResponseSchema": {"properties": {"monthly": {"type": "integer", "title": "Monthly"}, "total": {"type": "integer", "title": "Total"}, "limit": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Limit"}}, "type": "object", "required": ["monthly", "total", "limit"], "title": "GenerationCountResponseSchema"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MySubscriptionResponse": {"properties": {"plan": {"type": "string", "title": "Plan"}, "current_period_end": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Period End"}, "cancel_at_period_end": {"type": "boolean", "title": "Cancel At Period End"}, "has_stripe_customer": {"type": "boolean", "title": "Has Stripe Customer"}}, "type": "object", "required": ["plan", "current_period_end", "cancel_at_period_end", "has_stripe_customer"], "title": "MySubscriptionResponse"}, "PaginatedBlogPostResponseSchema": {"properties": {"items": {"items": {"$ref": "#/components/schemas/BlogPostResponseSchema"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total_pages": {"type": "integer", "title": "Total Pages"}}, "type": "object", "required": ["items", "total", "page", "page_size", "total_pages"], "title": "PaginatedBlogPostResponseSchema"}, "RagImageItemSchema": {"properties": {"image_url": {"type": "string", "title": "Image Url"}, "caption": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Caption"}, "page_number": {"type": "integer", "title": "Page Number"}}, "type": "object", "required": ["image_url", "caption", "page_number"], "title": "RagImageItemSchema"}, "RagSearchImageResponseSchema": {"properties": {"items": {"items": {"$ref": "#/components/schemas/RagImageItemSchema"}, "type": "array", "title": "Items"}}, "type": "object", "required": ["items"], "title": "RagSearchImageResponseSchema"}, "RagSearchRequestSchema": {"properties": {"query": {"type": "string", "minLength": 1, "title": "Query"}, "limit": {"type": "integer", "maximum": 50.0, "minimum": 1.0, "title": "Limit", "default": 5}}, "type": "object", "required": ["query"], "title": "RagSearchRequestSchema"}, "RagSearchTextResponseSchema": {"properties": {"chunks": {"items": {"$ref": "#/components/schemas/RagTextChunkSchema"}, "type": "array", "title": "Chunks"}}, "type": "object", "required": ["chunks"], "title": "RagSearchTextResponseSchema"}, "RagTextChunkSchema": {"properties": {"text": {"type": "string", "title": "Text"}, "page_number": {"type": "integer", "title": "Page Number"}}, "type": "object", "required": ["text", "page_number"], "title": "RagTextChunkSchema"}, "TargetLanguage": {"type": "string", "enum": ["japanese"], "title": "TargetLanguage", "description": "The target language for translation."}, "TextBlock": {"properties": {"type": {"type": "string", "const": "text", "title": "Type"}, "text": {"type": "string", "title": "Text"}}, "type": "object", "required": ["type", "text"], "title": "TextBlock", "description": "Anthropic \u4e92\u63db\u306e text content block\u3002`type` \u306f OpenAPI \u3067 required \u306b\u3059\u308b\n\u305f\u3081\u6562\u3048\u3066\u30c7\u30d5\u30a9\u30eb\u30c8\u5024\u3092\u6301\u305f\u305b\u305a\u3001\u30b3\u30f3\u30b9\u30c8\u30e9\u30af\u30bf\u3067\u660e\u793a\u7684\u306b\u6307\u5b9a\u3059\u308b\u904b\u7528\u3068\u3059\u308b\u3002"}, "ToolResultBlock": {"properties": {"type": {"type": "string", "const": "tool_result", "title": "Type"}, "tool_use_id": {"type": "string", "title": "Tool Use Id"}, "name": {"type": "string", "title": "Name"}, "content": {"additionalProperties": true, "type": "object", "title": "Content"}, "is_error": {"type": "boolean", "title": "Is Error", "default": false}}, "type": "object", "required": ["type", "tool_use_id", "name", "content"], "title": "ToolResultBlock"}, "ToolUseBlock": {"properties": {"type": {"type": "string", "const": "tool_use", "title": "Type"}, "id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "input": {"additionalProperties": true, "type": "object", "title": "Input"}}, "type": "object", "required": ["type", "id", "name"], "title": "ToolUseBlock"}, "TranslateResponseSchema": {"properties": {"message": {"type": "string", "title": "Message"}, "translated_pdf_url": {"type": "string", "maxLength": 2083, "minLength": 1, "format": "uri", "title": "Translated Pdf Url"}, "translated_language": {"type": "string", "title": "Translated Language"}}, "type": "object", "required": ["message", "translated_pdf_url", "translated_language"], "title": "TranslateResponseSchema"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}